MET-80: Tidy up components
