+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/index.tsx
@@ -2,7 +2,7 @@ import React, {StrictMode} from 'react';
 import {HashRouter as Router, Route, Switch} from 'react-router-dom';
 import {ThemeProvider} from 'styled-components';
 import {akeneoTheme} from 'akeneomeasure/shared/theme';
-import {Index} from 'akeneomeasure/pages/index/Index';
+import {Index} from 'akeneomeasure/pages/Index';
 import {LegacyContext, LegacyContextValue} from 'akeneomeasure/context/legacy-context';
 import {TranslateContext, TranslateContextValue} from 'akeneomeasure/context/translate-context';
 import {UserContext, UserContextValue} from 'akeneomeasure/context/user-context';
