+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/pages/Index.tsx
@@ -1,4 +1,4 @@
-import React, {useContext, useEffect, useState} from 'react';
+import React, {useContext} from 'react';
 import styled from 'styled-components';
 import {PageHeader} from 'akeneomeasure/shared/components/PageHeader';
 import {PageContent} from 'akeneomeasure/shared/components/PageContent';
@@ -6,44 +6,16 @@ import {PimView} from 'akeneomeasure/shared/components/pim-view/PimView';
 import {Breadcrumb} from 'akeneomeasure/shared/components/Breadcrumb';
 import {BreadcrumbItem} from 'akeneomeasure/shared/components/BreadcrumbItem';
 import {TranslateContext} from 'akeneomeasure/context/translate-context';
-import {
-  MeasurementFamily,
-  getMeasurementFamilyLabel,
-  getStandardUnitLabel,
-} from 'akeneomeasure/model/measurement-family';
-import {UserContext} from 'akeneomeasure/context/user-context';
-import {ResultCount} from 'akeneomeasure/shared/components/result-count';
 import {akeneoTheme} from 'akeneomeasure/shared/theme';
-import {Search as SearchIcon} from 'akeneomeasure/shared/icons/search';
 import {MeasurementFamily as MeasurementFamilyIllustration} from 'akeneomeasure/shared/illustrations/measurement-family';
 import {HelperTitle, HelperText, Helper} from 'akeneomeasure/shared/components/helper';
 import {Link} from 'akeneomeasure/shared/components/link';
 import {NoDataSection, NoDataTitle, NoDataText} from 'akeneomeasure/shared/components/no-data';
+import {useMeasurementFamilies} from 'akeneomeasure/hooks/use-measurement-families';
+import {MeasurementFamilyRow} from 'akeneomeasure/pages/index/measurement-family-row';
+import {SearchBar} from 'akeneomeasure/shared/components/search-bar';
 
-const List = styled.div``;
-
-const SearchBar = styled.div`
-  display: flex;
-  justify-content: space-between;
-  border-bottom: 1px solid ${akeneoTheme.color.grey100};
-  padding: 13px 0;
-  margin: 20px 0;
-`;
-const SearchContainer = styled.div`
-  display: flex;
-  align-items: center;
-`;
-const SearchInput = styled.input`
-  border: none;
-  width: 180px;
-  margin-left: 5px;
-  color: ${akeneoTheme.color.grey120};
-  outline: none;
-
-  ::placeholder {
-    color: ${akeneoTheme.color.grey120};
-  }
-`;
+const Container = styled.div``;
 
 const Table = styled.table`
   width: 100%;
@@ -54,91 +26,15 @@ const Table = styled.table`
     width: 25%;
   }
 `;
+
 const TableHeader = styled.thead`
   tr {
     height: 43px;
     border-bottom: 1px solid ${akeneoTheme.color.grey120};
   }
 `;
+
 const TableBody = styled.tbody``;
-const Row = styled.tr`
-  height: 54px;
-`;
-const MeasurementFamilyLabelCell = styled.td`
-  color: ${akeneoTheme.color.purple100};
-  font-style: italic;
-  font-weight: bold;
-`;
-
-const fetchMeasurementFamilies = (): Promise<MeasurementFamily[]> =>
-  Promise.resolve([
-    {
-      code: 'AREA',
-      labels: {en_US: 'Area', fr_FR: 'Aire'},
-      standard_unit_code: 'SQUARE_METER',
-      units: [
-        {
-          code: 'SQUARE_METER',
-          labels: {en_US: 'Square meter', fr_FR: 'metre carre'},
-          symbol: 'm2',
-          convert_from_standard: [{operator: 'mul', value: '1'}],
-        },
-      ],
-    },
-    {
-      code: 'LENGTH',
-      labels: {en_US: 'Length', fr_FR: 'Longueur'},
-      standard_unit_code: 'KILOMETER',
-      units: [
-        {
-          code: 'KILOMETER',
-          labels: {en_US: 'Kilometre', fr_FR: 'Kilomètre'},
-          symbol: 'm2',
-          convert_from_standard: [{operator: 'mul', value: '1'}],
-        },
-      ],
-    },
-    {
-      code: 'OTHER',
-      labels: {en_US: 'Other', fr_FR: 'Autre'},
-      standard_unit_code: 'ANOTHER_ONE',
-      units: [
-        {
-          code: 'ANOTHER_ONE',
-          labels: {en_US: 'Other unit', fr_FR: 'Autre unité'},
-          symbol: 'm2',
-          convert_from_standard: [{operator: 'mul', value: '1'}],
-        },
-      ],
-    },
-  ]);
-
-const useMeasurementFamilies = () => {
-  const [measurementFamilies, setMeasurementFamilies] = useState<MeasurementFamily[] | null>(null);
-
-  useEffect(() => {
-    (async () => setMeasurementFamilies(await fetchMeasurementFamilies()))();
-  }, []);
-
-  return measurementFamilies;
-};
-
-type MeasurementFamilyRowProps = {
-  measurementFamily: MeasurementFamily;
-};
-
-const MeasurementFamilyRow = ({measurementFamily}: MeasurementFamilyRowProps) => {
-  const locale = useContext(UserContext)('uiLocale');
-
-  return (
-    <Row>
-      <MeasurementFamilyLabelCell>{getMeasurementFamilyLabel(measurementFamily, locale)}</MeasurementFamilyLabelCell>
-      <td>{measurementFamily.code}</td>
-      <td>{getStandardUnitLabel(measurementFamily, locale)}</td>
-      <td>{measurementFamily.units.length}</td>
-    </Row>
-  );
-};
 
 export const Index = () => {
   const __ = useContext(TranslateContext);
@@ -196,14 +92,8 @@ export const Index = () => {
             </NoDataText>
           </NoDataSection>
         ) : (
-          <List>
-            <SearchBar>
-              <SearchContainer>
-                <SearchIcon />
-                <SearchInput placeholder={__('measurements.search.placeholder')} />
-              </SearchContainer>
-              <ResultCount count={measurementFamilies ? measurementFamilies.length : 0} />
-            </SearchBar>
+          <Container>
+            <SearchBar count={null === measurementFamilies ? null : measurementFamilies.length} />
             <Table>
               <TableHeader>
                 <tr>
@@ -219,7 +109,7 @@ export const Index = () => {
                 ))}
               </TableBody>
             </Table>
-          </List>
+          </Container>
         )}
       </PageContent>
     </>
